% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_aniview.R
\name{html_dependencies_aniview}
\alias{html_dependencies_aniview}
\title{AniView dependencies}
\usage{
html_dependencies_aniview(use_cdn = TRUE)
}
\arguments{
\item{use_cdn}{use CDN (by default) or use locally stored files.}
}
\description{
Get aniview html dependencies.
}
