% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniview.R
\name{aniview}
\alias{aniview}
\title{Animate an element}
\usage{
aniview(element, animation, ...)
}
\arguments{
\item{element}{An shiny or rmarkdown element.}

\item{animation}{An animation from \href{https://daneden.github.io/animate.css/}{animate.css}}

\item{...}{Additional class element}
}
\description{
Animate an element
}
